+++ packages/components/src/components/basic/typography/headline/headline-styles.spec.ts
@@ -2,42 +2,23 @@ import { getComponentCss, getSlottedCss } from './headline-styles';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    ['headline-1', 'left', 'default', false, 'light', undefined],
-    ['inherit', 'left', 'default', false, 'light', undefined],
-    ['large-title', 'center', 'inherit', true, 'dark', undefined],
-    ['headline-2', 'center', 'inherit', true, 'dark', undefined],
-    ['headline-3', 'center', 'inherit', true, 'dark', undefined],
-    ['headline-4', 'center', 'inherit', true, 'dark', undefined],
-    ['headline-5', 'center', 'inherit', true, 'dark', undefined],
+    ['headline-1', 'left', 'default', false, 'light'],
+    ['inherit', 'left', 'default', false, 'light'],
+    ['large-title', 'center', 'inherit', true, 'dark'],
+    ['headline-2', 'center', 'inherit', true, 'dark'],
+    ['headline-3', 'center', 'inherit', true, 'dark'],
+    ['headline-4', 'center', 'inherit', true, 'dark'],
+    ['headline-5', 'center', 'inherit', true, 'dark'],
     [
       { base: 'small', xs: 'large', s: 'small', m: 'large', l: 'small', xl: 'large' },
       'right',
       'inherit',
       false,
       'dark',
-      undefined,
     ],
-    ['headline-1', 'left', 'default', false, 'light', 'none'],
-    ['inherit', 'left', 'default', false, 'light', 'none'],
-    ['large-title', 'center', 'inherit', true, 'dark', 'none'],
-    ['headline-2', 'center', 'inherit', true, 'dark', 'none'],
-    ['headline-3', 'center', 'inherit', true, 'dark', 'none'],
-    ['headline-4', 'center', 'inherit', true, 'dark', 'none'],
-    ['headline-5', 'center', 'inherit', true, 'dark', 'none'],
-    [
-      { base: 'small', xs: 'large', s: 'small', m: 'large', l: 'small', xl: 'large' },
-      'right',
-      'inherit',
-      false,
-      'dark',
-      'none',
-    ],
-  ])(
-    'should return correct css for variant: %j, align: %s, color: %s, ellipsis: %s, theme: %s and hyphens: %s',
-    (...args) => {
-      expect(getComponentCss(...args)).toMatchSnapshot();
-    }
-  );
+  ])('should return correct css for variant: %j, align: %s, color: %s, ellipsis: %s and theme: %s', (...args) => {
+    expect(getComponentCss(...args)).toMatchSnapshot();
+  });
 });
 
 describe('getSlottedCss()', () => {