+++ packages/components/CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-#### Added
-- `Headline` `hyphens` property which can be set to `none`
+#### Changed
+- `Headline` `hyphens` style to `manual`
 
 ### [2.9.3-rc.0-skeletons] - 2022-03-29